Bring shell into foreground if in session leader's process group
This commit implements part of the initialization of the job-control
shell that is specified by POSIX.1-2024, which says that the shell
should bring itself into the foreground only if it is started as the
session leader for the terminal session. However, this commit also
brings the shell into the foreground if the shell is in the same
process group as the session leader because it is unlikely that there
is another job-controlling process that can bring the shell into the
foreground.

This commit changes the behavior of the shell so that the shell is
brought into the foreground if it is started in the same process group
as the session leader, rather than suspending the shell until it is
brought to the foreground by some other means.

Closes https://github.com/magicant/yash/issues/134