Avoid flakiness in job suspension test
The test case 'discard remaining commands when a command suspends' in
tests/job-y.tst fails very rarely. It seems that the echo command may
get suspended after the second pipeline component is resumed for the
second time when the SIGSTOP signal is delivered so late. This patch
avoids the flakiness by removing the pipeline component so that there is
only one direct child process of the shell in the job.