Rename E_SUSPEND to E_CANCEL
This commit renames the E_SUSPEND exception to E_CANCEL. The E_CANCEL
exception will be raised not only when a foreground job is suspended but
also when a shell error occurs that would terminate the shell if it were
not interactive. This change is made to make the exception name match
its purpose.