Return exit status 128 on command read error
POSIX.1-2024 requires that the shell return exit status 128 when
it encounters an unrecoverable read error, except for the dot built-in
which should return an exit status between 1 and 125.
This commit changes the exit status to 128 for unrecoverable read
errors while keeping the exit status for the dot built-in at 2.

See https://www.austingroupbugs.net/view.php?id=1629