Rewrite tests for . and .. in pathname expansion
While implementing the yash-rs equivalent of commit e231437f, I noticed
that the behavior around the filtering of . and .. in pathname expansion
was not actually relevant to the extendedglob feature. This commit
rewrites the tests so that they do not depend on extendedglob. The test
case 'literal . and .. are not filtered out' is also moved to path-p.tst
as it is mandatory in POSIX.

This commit re-imports the test files from:
https://github.com/magicant/yash-rs/commit/5863fe2f104cdb249989886d8da3e9eaa0597036