Reject multiple pathname expansion results in redirection operands
POSIX.1-2024 allows the shell to treat it as an error when pathname
expansion on a file redirection operand results in more than one
pathname. Previously, yash only did so when the POSIXly-correct mode was
disabled, but now it does so regardless of the mode.

POSIX does not seem to allow the shell to treat it as an error when
pathname expansion on a file redirection operand results in no pathname,
so this error is only reported when the POSIXly-correct mode is disabled
as before. However, the previous error message was misleading as we were
using the same message as the one for multiple matches. This commit
changes the error message to indicate that no match was found.

Closes https://github.com/magicant/yash/issues/95