Fix default exit status of exit built-in in trap action
This commit fixes the behavior of the `exit` built-in when used in a
trap action without an operand. Previously, the shell would exit with
the exit status of the last command executed in the trap action. This
commit changes it to exit with the exit status of the last command
executed before the trap action, as per POSIX specifications.