cd: Return exit status 1 if $PWD or $OLDPWD is read-only
In section XBD 8.1 of POSIX.1-2024, the expected behavior of the shell
is defined for the case when the `readonly` built-in is executed on the
`$PWD` or `$OLDPWD` variable. The standard allows us to choose from
three options, and we make the `cd` (and `pushd`) built-ins fail with
non-zero exit status in this case.