Optimize array built-in
The previous implementation of array_remove_elements called pl_remove
for each element to be removed, which had a time complexity of O(m*n)
where m is the number of elements to be removed and n is the number of
elements in the array. The new implementation moves each element at most
once, resulting in a time complexity of O(max(m,n)).

This commit also shortcuts a call to update_environment if no elements
are modified.