parser: No single quotes in subst in double-quoted parameter
If a parameter expansion with a PT_MINUS, PT_PLUS, PT_ASSIGN, or
PT_ERROR type is in double quotes, single quotes are not treated as
quoting characters in the substitution word. This is what is specified
in POSIX.1-2024 XCU 2.2.3:

> For parameter expansions other than the four varieties that provide
> for substring processing, within the string of characters from an
> enclosed "${" to the matching '}', the double-quotes within which the
> expansion occurs shall preserve the literal value of all characters,
> with the exception of the characters double-quote, backquote,
> <dollar-sign>, and <backslash>.

Commit a333044c attempted to fix this, but it only addressed the
expansion behavior. This commit fixes the parsing behavior.

The parser needs to be aware of whether the parameter expansion is in
double quotes or not, so the `parse_paramexp_in_brace` function now
takes an additional `indq` argument to indicate if the expansion is in
double quotes. This flag is passed down to the `parse_word` function,
which then decides whether to allow single quotes in the substitution
word.

Fixes https://github.com/magicant/yash/issues/174