No tilde expansion in subst in double-quoted parameter
If a parameter expansion with a PT_MINUS, PT_PLUS, PT_ASSIGN, or
PT_ERROR type is in double quotes, tilde expansion is not performed
on the substitution part of the expansion. This is what is specified
in POSIX.1-2024 XCU 2.2.3:

> For parameter expansions other than the four varieties that provide
> for substring processing, within the string of characters from an
> enclosed "${" to the matching '}', the double-quotes within which the
> expansion occurs shall preserve the literal value of all characters,
> with the exception of the characters double-quote, backquote,
> <dollar-sign>, and <backslash>.

Fixes https://github.com/magicant/yash/issues/175

See also commit 4eccc94c, which addressed a similar issue for single
quotes.